pyproject.toml
[tool.poetry]
name = "google-calendar-analytics"
version = "0.5.0"
description = "A Python library for analyzing Google Calendar data."
authors = [
    "Berupor <evgeniy.zelenoff@gmail.com>"
]
readme = "README.md"
packages = [{ include = "google_calendar_analytics" }]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: OS Independent",
]
keywords = [
    "google",
    "calendar",
    "analytics",
    "data analysis",
    "event tracking",
    "attendee tracking",
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "1.5.3"
google-auth = "2.16.0"
google-api-python-client = "2.77.0"
google-auth-httplib2 = "0.1.0"
google-auth-oauthlib = "1.0.0"
plotly = "5.13.0"
kaleido = "0.2.1"
aiohttp = "3.8.4"
certifi = "2022.12.7"

[tool.poetry.group.dev.dependencies]
mypy = "1.0.1"
flake8 = "6.0.0"
black = "*"
isort = "*"
pytest = "7.2.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "1.4.2"
mkdocs-extensions = "0.1.2"
mkdocstrings = "0.20.0"
mkdocs-material = "9.1.3"

[project.urls]
Documentation = "https://berupor.github.io/Calendar-Analytics/"
Repository = "https://github.com/Berupor/Calendar-Analytics"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"